fix: preserve decimals in values (penetration, avg price, frequency, ...) (#7)
Decimal KPIs were shown as integers ('46' instead of 45.9; '7' for 7.x),
because to_key_measures() rounded floats and the answer formatters used .0f.

- pivot_result.format_number/format_plain: adaptive formatting that keeps up to
  2 decimals, trims trailing zeros, and groups thousands (chat) or not (CSV).
- answer_from_pivot_tables: format per metric type — signed % for Yr-on-Yr
  change, unsigned % for level percentages (Penetration %), plain decimals
  otherwise. Pivot answers now always come from the float result table.
- to_key_measures no longer rounds; KeyMeasuresTable values may be float, so
  CSV export and the data checker keep decimals too.

Verified live: 金果 平均价格 45.9 (was 46); 销额 1,017,973. Tests: 111 passed.

+++ app/worldpanel/parser.py
@@ -14,9 +14,10 @@ class KeyMeasuresTable:
     metric: str
     products: list[str]
     dates: list[str]
-    rows: dict[str, dict[str, int]]
+    # Values may be int (Spend) or float (Penetration %, Average Price, ...).
+    rows: dict[str, dict[str, float]]
 
-    def value_for(self, product: str, date_label: str, metric: str | None = None) -> int:
+    def value_for(self, product: str, date_label: str, metric: str | None = None) -> float:
         if metric and metric.lower() != self.metric.lower():
             raise KeyError(f"Metric not available: {metric}")
         canonical_product = _match_product(product, self.products)
